+++ src/Statistics.hpp
@@ -15,16 +15,39 @@
 
 using namespace std;
 
+/**
+ * Get statistic of correct frames' number.
+ *
+ * Contains information about correctness of read frames.
+ */
 class Statistics {
 private:
-    int correct;
-    int incorrect;
-    unsigned long long whole;
+    /// Numer of correct frames.
+    u_int64_t correct;
+    /// Number of incorrect frames.
+    u_int64_t incorrect;
+    /// Total number of frames.
+    u_int64_t whole;
 public:
-    Statistics(int whole_v, int correct_v, int incorrect_v);
-    
+    /**
+     * Basic constructor.
+     *
+     * @param whole_v total number of frames.
+     * @param correct_v numer of correct frames.
+     * @param incorrect_v number of incorrect frames.
+     */
+    Statistics(u_int64_t whole_v, u_int64_t correct_v, u_int64_t incorrect_v);
+    /**
+     * Alternative constructor.
+     *
+     * @param frames reference to vector of read frames.
+     */
     Statistics(const vector<LogFrame> &frames);
-    
+    /**
+     * Convert statistic to string value.
+     *
+     * @return string representation of statistic.
+     */
     string toString();
 };
